in reference to prev docs, see in Tilemap.js
"A Tilemap is a container for Tilemap data. This isn't a display object, rather, it holds data about the map and allows you to add tilesets and tilemap layers to it. A map can have one or more tilemap layers, which are the display objects that actually render the tiles."